Implement import and export using module.
