+++ module/renderBook.js
@@ -1,15 +1,25 @@
-const titleInputEl = document.getElementById('title-input');
-const authorInputEl = document.getElementById('author-input');
-const formEl = document.getElementById('form');
+const renderBooks = (collectionsSectionEl, bookCollection) => {
+  collectionsSectionEl.innerHTML = '';
+  bookCollection.collectionData.forEach((book, index) => {
+    const bookEntry = document.createElement('div');
+    bookEntry.classList.add('book-entry');
+    const bookInfo = document.createElement('span');
+    bookInfo.classList.add('book-info');
+    bookInfo.textContent = `"${book.title}" by ${book.author}`;
+    bookCollection.removeButtons[index] = document.createElement('button');
+    bookCollection.removeButtons[index].classList.add('remove-button');
+    bookCollection.removeButtons[index].textContent = 'Remove';
+    bookCollection.removeButtons[index].dataset.id = book.bookID;
+    bookEntry.append(bookInfo, bookCollection.removeButtons[index]);
+    collectionsSectionEl.appendChild(bookEntry);
+  });
 
-formEl.addEventListener('submit', (e) => {
-  e.preventDefault();
-  const title = titleInputEl.value;
-  const author = authorInputEl.value;
-  const removeIndex = bookCollection.collectionData.length;
-  const book = { title, author, bookID: removeIndex };
-  bookCollection.addBook(book);
-  renderBooks(bookCollection.collectionData);
-});
+  bookCollection.removeButtons.forEach((button) => {
+    button.addEventListener('click', (e) => {
+      bookCollection.removeBook(+e.target.dataset.id);
+      renderBooks(collectionsSectionEl, bookCollection);
+    });
+  });
+};
 
-export default addbook
+export default renderBooks;
